Make all purge-plugin dataformats header-only, revert libXrdPfc lib type to MODULE, cleanup some includes.
