+++ src/XrdPfc/XrdPfcDirStateSnapshot.hh
@@ -15,8 +15,6 @@ class XrdOss;
 namespace XrdPfc
 {
 
-// For usage / stat reporting
-
 struct DirStateElement : public DirStateBase
 {
    DirStats m_stats;
@@ -40,9 +38,9 @@ struct DataFsSnapshot : public DataFsStateBase
    time_t                       m_sshot_stats_reset_time = 0;
 
    DataFsSnapshot() {}
-   DataFsSnapshot(const DataFsState &b) :
+   DataFsSnapshot(const DataFsStateBase &b, time_t sshot_stats_reset_time) :
      DataFsStateBase(b),
-     m_sshot_stats_reset_time(b.m_sshot_stats_reset_time)
+     m_sshot_stats_reset_time(sshot_stats_reset_time)
    {}
 
    // Import of data into vector form is implemented in ResourceMonitor
@@ -50,48 +48,6 @@ struct DataFsSnapshot : public DataFsStateBase
 
    void write_json_file(const std::string &fname, XrdOss& oss, bool include_preamble);
    void dump();
-};
-
-// For purge planning & execution
-
-struct DirPurgeElement : public DirStateBase
-{
-   DirUsage m_usage;
-
-   int m_parent = -1;
-   int m_daughters_begin = -1, m_daughters_end = -1;
-
-   DirPurgeElement() {}
-   DirPurgeElement(const DirState &b, int parent) :
-     DirStateBase(b),
-     m_usage(b.m_here_usage, b.m_recursive_subdir_usage),
-     m_parent(parent)
-   {}
-};
-
-struct DataFsPurgeshot : public DataFsStateBase
-{
-   long long m_bytes_to_remove = 0;
-   long long m_estimated_writes_from_writeq = 0;
-
-   bool m_space_based_purge = false;
-   bool m_age_based_purge = false;
-
-   std::vector<DirPurgeElement> m_dir_vec;
-   // could have parallel vector of DirState* ... or store them in the DirPurgeElement.
-   // requires some interlock / ref-counting with the source tree.
-   // or .... just block DirState removal for the duration of the purge :) Yay.
-
-   DataFsPurgeshot() {}
-   DataFsPurgeshot(const DataFsState &b) :
-     DataFsStateBase(b)
-   {}
-
-  int find_dir_entry_from_tok(int entry, PathTokenizer &pt, int pos, int *last_existing_entry) const;
-
-  int find_dir_entry_for_dir_path(const std::string &dir_path) const;
-
-  const DirUsage* find_dir_usage_for_dir_path(const std::string &dir_path) const;
 };
 
 }
